Refactoring + shuffling & sampling (#1)
* refactoring trial

* fix iter

* add split datasets, remove iter sample episodes

* add EpisodeData

* address Rodrigo's comments

* add spec

* add total_steps initialization
